split converter into core and utilities
